+++ pybamm/__init__.py
@@ -266,6 +266,8 @@ from .solvers.casadi_solver import CasadiSolver
 from .solvers.scikits_dae_solver import ScikitsDaeSolver
 from .solvers.scikits_ode_solver import ScikitsOdeSolver, have_scikits_odes
 from .solvers.scipy_solver import ScipySolver
+from .solvers.klu_sparse_solver import KLU, have_klu
+
 
 #
 # Current profiles